address pr feedback, fix namespace leak filter before unprefixing, added coverage. cleaned up large executor file
