+++ pkg/cypher/executor.go
@@ -115,14 +115,12 @@ import (
 	"fmt"
 	"io"
 	"log/slog"
-	"strconv"
 	"strings"
 	"sync"
 	"time"
 	"unicode"
 
 	"github.com/orneryd/nornicdb/pkg/config"
-	"github.com/orneryd/nornicdb/pkg/cypher/antlr"
 	"github.com/orneryd/nornicdb/pkg/embeddingutil"
 	nornicerrors "github.com/orneryd/nornicdb/pkg/errors"
 	"github.com/orneryd/nornicdb/pkg/fabric"
@@ -2357,1396 +2355,4 @@ func (e *StorageExecutor) resolveWALAndDatabase() (*storage.WAL, string) {
 	}
 
 	return nil, dbName
-}
-
-// transactionStorageWrapper wraps a BadgerTransaction to implement storage.Engine
-// for use in implicit transaction execution. It routes writes through the transaction
-// (for atomicity/rollback) and reads through the underlying engine (for performance).
-type transactionStorageWrapper struct {
-	tx               *storage.BadgerTransaction
-	underlying       storage.Engine // For read operations not supported by transaction
-	namespace        string
-	separator        string
-	mutatedNodeIDs   map[string]struct{}
-	mutatedNodeIDsMu sync.Mutex
-
-	// txNodeLookupCache scopes the executor's MERGE/MATCH lookup cache to
-	// this single transaction. Concurrent transactions get distinct
-	// wrappers — and therefore distinct caches — so a peer's uncommitted
-	// node-ID mapping cannot leak into this transaction's read set via
-	// store.GetNode(...) inside tx.badgerTx (which would otherwise put
-	// the peer's node key into Badger's SSI read set and convert a
-	// constraint violation into a generic Transaction Conflict). Within
-	// a single transaction the cache is shared across re-entries that
-	// reuse the same wrapper, so multi-clause queries still benefit
-	// from the cross-clause speedup.
-	txNodeLookupCache   map[string]*storage.Node
-	txNodeLookupCacheMu *sync.RWMutex
-}
-
-func (w *transactionStorageWrapper) Namespace() string {
-	if w == nil {
-		return ""
-	}
-	return w.namespace
-}
-
-// ensureNodeLookupCacheLocked lazily initializes the wrapper's MERGE/MATCH
-// lookup cache, seeding it once from the parent executor's committed
-// entries. The cache exists for the lifetime of the transaction; on
-// commit, executor code drains it back into the parent executor via
-// promoteNodeLookupCacheTo, on rollback the wrapper is discarded and the
-// cache with it. Subsequent calls (e.g. recursive Execute re-entry on
-// the same wrapper) are no-ops, so the in-tx state survives across
-// multi-clause queries.
-func (w *transactionStorageWrapper) ensureNodeLookupCacheLocked(seedFrom *StorageExecutor) {
-	if w.txNodeLookupCacheMu != nil && w.txNodeLookupCache != nil {
-		return
-	}
-	w.txNodeLookupCacheMu = &sync.RWMutex{}
-	w.txNodeLookupCache = make(map[string]*storage.Node, 1000)
-	if seedFrom == nil {
-		return
-	}
-	srcMu := seedFrom.nodeLookupCacheLock()
-	srcMu.RLock()
-	for k, v := range seedFrom.nodeLookupCache {
-		w.txNodeLookupCache[k] = v
-	}
-	srcMu.RUnlock()
-}
-
-func (w *transactionStorageWrapper) GetEngine() storage.Engine {
-	if w == nil {
-		return nil
-	}
-	return w.underlying
-}
-
-func (w *transactionStorageWrapper) markMutatedNodeID(id storage.NodeID) {
-	if id == "" || w.mutatedNodeIDs == nil {
-		return
-	}
-	w.mutatedNodeIDsMu.Lock()
-	w.mutatedNodeIDs[string(id)] = struct{}{}
-	w.mutatedNodeIDsMu.Unlock()
-}
-
-func (w *transactionStorageWrapper) snapshotMutatedNodeIDs() map[string]struct{} {
-	if len(w.mutatedNodeIDs) == 0 {
-		return nil
-	}
-	w.mutatedNodeIDsMu.Lock()
-	defer w.mutatedNodeIDsMu.Unlock()
-	out := make(map[string]struct{}, len(w.mutatedNodeIDs))
-	for id := range w.mutatedNodeIDs {
-		out[id] = struct{}{}
-	}
-	return out
-}
-
-// Write operations - go through transaction for atomicity
-func (w *transactionStorageWrapper) CreateNode(node *storage.Node) (storage.NodeID, error) {
-	if w.namespace == "" {
-		id, err := w.tx.CreateNode(node)
-		if err == nil {
-			w.markMutatedNodeID(id)
-		}
-		return id, err
-	}
-	namespaced := storage.CopyNode(node)
-	namespaced.ID = w.prefixNodeID(node.ID)
-	actualID, err := w.tx.CreateNode(namespaced)
-	if err != nil {
-		return "", err
-	}
-	userID := w.unprefixNodeID(actualID)
-	w.markMutatedNodeID(userID)
-	return userID, nil
-}
-
-func (w *transactionStorageWrapper) UpdateNode(node *storage.Node) error {
-	if w.namespace == "" {
-		err := w.tx.UpdateNode(node)
-		if err == nil {
-			w.markMutatedNodeID(node.ID)
-		}
-		return err
-	}
-	namespaced := storage.CopyNode(node)
-	namespaced.ID = w.prefixNodeID(node.ID)
-	err := w.tx.UpdateNode(namespaced)
-	if err == nil {
-		w.markMutatedNodeID(node.ID)
-	}
-	return err
-}
-
-func (w *transactionStorageWrapper) DeleteNode(id storage.NodeID) error {
-	return w.tx.DeleteNode(w.prefixNodeID(id))
-}
-
-func (w *transactionStorageWrapper) CreateEdge(edge *storage.Edge) error {
-	if w.namespace == "" {
-		return w.tx.CreateEdge(edge)
-	}
-	namespaced := storage.CopyEdge(edge)
-	namespaced.ID = w.prefixEdgeID(edge.ID)
-	namespaced.StartNode = w.prefixNodeID(edge.StartNode)
-	namespaced.EndNode = w.prefixNodeID(edge.EndNode)
-	return w.tx.CreateEdge(namespaced)
-}
-
-func (w *transactionStorageWrapper) DeleteEdge(id storage.EdgeID) error {
-	return w.tx.DeleteEdge(w.prefixEdgeID(id))
-}
-
-// Read operations - transaction supports GetNode, forward others to underlying
-func (w *transactionStorageWrapper) GetNode(id storage.NodeID) (*storage.Node, error) {
-	node, err := w.tx.GetNode(w.prefixNodeID(id))
-	if err != nil {
-		return nil, err
-	}
-	if w.namespace == "" {
-		return node, nil
-	}
-	return w.toUserNode(node), nil
-}
-
-func (w *transactionStorageWrapper) GetEdge(id storage.EdgeID) (*storage.Edge, error) {
-	if w.namespace == "" {
-		return w.tx.GetEdge(id)
-	}
-	edge, err := w.tx.GetEdge(w.prefixEdgeID(id))
-	if err != nil {
-		return nil, err
-	}
-	return w.toUserEdge(edge), nil
-}
-
-func (w *transactionStorageWrapper) UpdateEdge(edge *storage.Edge) error {
-	if w.namespace == "" {
-		return w.tx.UpdateEdge(edge)
-	}
-	namespaced := storage.CopyEdge(edge)
-	namespaced.ID = w.prefixEdgeID(edge.ID)
-	namespaced.StartNode = w.prefixNodeID(edge.StartNode)
-	namespaced.EndNode = w.prefixNodeID(edge.EndNode)
-	return w.tx.UpdateEdge(namespaced)
-}
-
-func (w *transactionStorageWrapper) GetNodesByLabel(label string) ([]*storage.Node, error) {
-	nodes, err := w.tx.GetNodesByLabel(label)
-	if err != nil {
-		return nil, err
-	}
-	if w.namespace == "" {
-		return nodes, nil
-	}
-	return w.toUserNodes(nodes), nil
-}
-
-func (w *transactionStorageWrapper) GetFirstNodeByLabel(label string) (*storage.Node, error) {
-	node, err := w.tx.GetFirstNodeByLabel(label)
-	if err != nil {
-		return nil, err
-	}
-	if w.namespace == "" {
-		return node, nil
-	}
-	return w.toUserNode(node), nil
-}
-
-func (w *transactionStorageWrapper) ForEachNodeIDByLabel(label string, visit func(storage.NodeID) bool) error {
-	if visit == nil {
-		return nil
-	}
-	if !w.tx.HasPendingNodeMutations() {
-		if lookup, ok := w.underlying.(storage.LabelNodeIDLookupEngine); ok {
-			return lookup.ForEachNodeIDByLabel(label, visit)
-		}
-	}
-
-	nodes, err := w.GetNodesByLabel(label)
-	if err != nil {
-		return err
-	}
-	for _, node := range nodes {
-		if node == nil {
-			continue
-		}
-		if !visit(node.ID) {
-			return nil
-		}
-	}
-	return nil
-}
-
-func (w *transactionStorageWrapper) GetOutgoingEdges(nodeID storage.NodeID) ([]*storage.Edge, error) {
-	if w.namespace == "" {
-		return w.tx.GetOutgoingEdges(nodeID)
-	}
-	edges, err := w.tx.GetOutgoingEdges(w.prefixNodeID(nodeID))
-	if err != nil {
-		return nil, err
-	}
-	return w.toUserEdges(edges), nil
-}
-
-func (w *transactionStorageWrapper) GetIncomingEdges(nodeID storage.NodeID) ([]*storage.Edge, error) {
-	if w.namespace == "" {
-		return w.tx.GetIncomingEdges(nodeID)
-	}
-	edges, err := w.tx.GetIncomingEdges(w.prefixNodeID(nodeID))
-	if err != nil {
-		return nil, err
-	}
-	return w.toUserEdges(edges), nil
-}
-
-func (w *transactionStorageWrapper) GetEdgesBetween(startID, endID storage.NodeID) ([]*storage.Edge, error) {
-	if w.namespace == "" {
-		return w.tx.GetEdgesBetween(startID, endID)
-	}
-	edges, err := w.tx.GetEdgesBetween(w.prefixNodeID(startID), w.prefixNodeID(endID))
-	if err != nil {
-		return nil, err
-	}
-	return w.toUserEdges(edges), nil
-}
-
-func (w *transactionStorageWrapper) GetEdgeBetween(startID, endID storage.NodeID, edgeType string) *storage.Edge {
-	if w.namespace == "" {
-		return w.tx.GetEdgeBetween(startID, endID, edgeType)
-	}
-	edge := w.tx.GetEdgeBetween(w.prefixNodeID(startID), w.prefixNodeID(endID), edgeType)
-	if edge == nil {
-		return nil
-	}
-	return w.toUserEdge(edge)
-}
-
-func (w *transactionStorageWrapper) GetEdgesByType(edgeType string) ([]*storage.Edge, error) {
-	if w.namespace == "" {
-		return w.tx.GetEdgesByType(edgeType)
-	}
-	edges, err := w.tx.GetEdgesByType(edgeType)
-	if err != nil {
-		return nil, err
-	}
-	return w.toUserEdges(edges), nil
-}
-
-func (w *transactionStorageWrapper) GetNodesByLabelVisibleAt(label string, version storage.MVCCVersion) ([]*storage.Node, error) {
-	if provider, ok := w.underlying.(storage.MVCCIndexedVisibilityEngine); ok {
-		return provider.GetNodesByLabelVisibleAt(label, version)
-	}
-	return nil, storage.ErrNotImplemented
-}
-
-func (w *transactionStorageWrapper) GetOutgoingEdgesVisibleAt(nodeID storage.NodeID, version storage.MVCCVersion) ([]*storage.Edge, error) {
-	provider, ok := w.underlying.(storage.MVCCIndexedVisibilityEngine)
-	if !ok {
-		return nil, storage.ErrNotImplemented
-	}
-	edges, err := provider.GetOutgoingEdgesVisibleAt(w.prefixNodeID(nodeID), version)
-	if err != nil || w.namespace == "" {
-		return edges, err
-	}
-	return w.toUserEdges(edges), nil
-}
-
-func (w *transactionStorageWrapper) GetIncomingEdgesVisibleAt(nodeID storage.NodeID, version storage.MVCCVersion) ([]*storage.Edge, error) {
-	provider, ok := w.underlying.(storage.MVCCIndexedVisibilityEngine)
-	if !ok {
-		return nil, storage.ErrNotImplemented
-	}
-	edges, err := provider.GetIncomingEdgesVisibleAt(w.prefixNodeID(nodeID), version)
-	if err != nil || w.namespace == "" {
-		return edges, err
-	}
-	return w.toUserEdges(edges), nil
-}
-
-func (w *transactionStorageWrapper) GetEdgesByTypeVisibleAt(edgeType string, version storage.MVCCVersion) ([]*storage.Edge, error) {
-	if provider, ok := w.underlying.(storage.MVCCIndexedVisibilityEngine); ok {
-		return provider.GetEdgesByTypeVisibleAt(edgeType, version)
-	}
-	return nil, storage.ErrNotImplemented
-}
-
-func (w *transactionStorageWrapper) GetEdgesBetweenVisibleAt(startID, endID storage.NodeID, version storage.MVCCVersion) ([]*storage.Edge, error) {
-	if provider, ok := w.underlying.(storage.MVCCIndexedVisibilityEngine); ok {
-		return provider.GetEdgesBetweenVisibleAt(startID, endID, version)
-	}
-	return nil, storage.ErrNotImplemented
-}
-
-func (w *transactionStorageWrapper) AllNodes() ([]*storage.Node, error) {
-	nodes, err := w.tx.AllNodes()
-	if err != nil {
-		return nil, err
-	}
-	if w.namespace == "" {
-		return nodes, nil
-	}
-	return w.toUserNodes(nodes), nil
-}
-
-func (w *transactionStorageWrapper) AllEdges() ([]*storage.Edge, error) {
-	return w.underlying.AllEdges()
-}
-
-func (w *transactionStorageWrapper) GetAllNodes() []*storage.Node {
-	nodes := w.tx.GetAllNodes()
-	if w.namespace == "" {
-		return nodes
-	}
-	return w.toUserNodes(nodes)
-}
-
-func (w *transactionStorageWrapper) GetInDegree(nodeID storage.NodeID) int {
-	return w.underlying.GetInDegree(nodeID)
-}
-
-func (w *transactionStorageWrapper) GetOutDegree(nodeID storage.NodeID) int {
-	return w.underlying.GetOutDegree(nodeID)
-}
-
-func (w *transactionStorageWrapper) GetSchema() *storage.SchemaManager {
-	return w.underlying.GetSchema()
-}
-
-func (w *transactionStorageWrapper) BulkCreateNodes(nodes []*storage.Node) error {
-	// For bulk operations within transaction, create one by one
-	for _, node := range nodes {
-		if w.namespace == "" {
-			if _, err := w.tx.CreateNode(node); err != nil {
-				return err
-			}
-			continue
-		}
-		namespaced := storage.CopyNode(node)
-		namespaced.ID = w.prefixNodeID(node.ID)
-		if _, err := w.tx.CreateNode(namespaced); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (w *transactionStorageWrapper) BulkCreateEdges(edges []*storage.Edge) error {
-	if len(edges) == 0 {
-		return nil
-	}
-	if w.namespace == "" {
-		return w.tx.BulkCreateEdges(edges)
-	}
-	namespaced := make([]*storage.Edge, len(edges))
-	for i, edge := range edges {
-		cp := storage.CopyEdge(edge)
-		cp.ID = w.prefixEdgeID(edge.ID)
-		cp.StartNode = w.prefixNodeID(edge.StartNode)
-		cp.EndNode = w.prefixNodeID(edge.EndNode)
-		namespaced[i] = cp
-	}
-	return w.tx.BulkCreateEdges(namespaced)
-}
-
-func (w *transactionStorageWrapper) BulkDeleteNodes(ids []storage.NodeID) error {
-	for _, id := range ids {
-		if err := w.tx.DeleteNode(w.prefixNodeID(id)); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (w *transactionStorageWrapper) BulkDeleteEdges(ids []storage.EdgeID) error {
-	for _, id := range ids {
-		if err := w.tx.DeleteEdge(w.prefixEdgeID(id)); err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-func (w *transactionStorageWrapper) prefixNodeID(id storage.NodeID) storage.NodeID {
-	if w.namespace == "" {
-		return id
-	}
-	prefix := w.namespace + w.separator
-	if strings.HasPrefix(string(id), prefix) {
-		return id
-	}
-	return storage.NodeID(w.namespace + w.separator + string(id))
-}
-
-func (w *transactionStorageWrapper) unprefixNodeID(id storage.NodeID) storage.NodeID {
-	if w.namespace == "" {
-		return id
-	}
-	prefix := w.namespace + w.separator
-	s := string(id)
-	if strings.HasPrefix(s, prefix) {
-		return storage.NodeID(s[len(prefix):])
-	}
-	return id
-}
-
-func (w *transactionStorageWrapper) prefixEdgeID(id storage.EdgeID) storage.EdgeID {
-	if w.namespace == "" {
-		return id
-	}
-	prefix := w.namespace + w.separator
-	if strings.HasPrefix(string(id), prefix) {
-		return id
-	}
-	return storage.EdgeID(w.namespace + w.separator + string(id))
-}
-
-func (w *transactionStorageWrapper) unprefixEdgeID(id storage.EdgeID) storage.EdgeID {
-	if w.namespace == "" {
-		return id
-	}
-	prefix := w.namespace + w.separator
-	s := string(id)
-	if strings.HasPrefix(s, prefix) {
-		return storage.EdgeID(s[len(prefix):])
-	}
-	return id
-}
-
-func (w *transactionStorageWrapper) toUserNode(node *storage.Node) *storage.Node {
-	if node == nil {
-		return nil
-	}
-	out := storage.CopyNode(node)
-	out.ID = w.unprefixNodeID(out.ID)
-	return out
-}
-
-func (w *transactionStorageWrapper) toUserEdge(edge *storage.Edge) *storage.Edge {
-	if edge == nil {
-		return nil
-	}
-	out := storage.CopyEdge(edge)
-	out.ID = w.unprefixEdgeID(out.ID)
-	out.StartNode = w.unprefixNodeID(out.StartNode)
-	out.EndNode = w.unprefixNodeID(out.EndNode)
-	return out
-}
-
-func (w *transactionStorageWrapper) toUserEdges(edges []*storage.Edge) []*storage.Edge {
-	out := make([]*storage.Edge, 0, len(edges))
-	for _, edge := range edges {
-		out = append(out, w.toUserEdge(edge))
-	}
-	return out
-}
-
-func (w *transactionStorageWrapper) toUserNodes(nodes []*storage.Node) []*storage.Node {
-	out := make([]*storage.Node, 0, len(nodes))
-	for _, node := range nodes {
-		out = append(out, w.toUserNode(node))
-	}
-	return out
-}
-
-func (w *transactionStorageWrapper) BatchGetNodes(ids []storage.NodeID) (map[storage.NodeID]*storage.Node, error) {
-	return w.underlying.BatchGetNodes(ids)
-}
-
-func (w *transactionStorageWrapper) Close() error {
-	// Don't close underlying engine
-	return nil
-}
-
-func (w *transactionStorageWrapper) NodeCount() (int64, error) {
-	return w.underlying.NodeCount()
-}
-
-func (w *transactionStorageWrapper) EdgeCount() (int64, error) {
-	return w.underlying.EdgeCount()
-}
-
-func (w *transactionStorageWrapper) DeleteByPrefix(prefix string) (nodesDeleted int64, edgesDeleted int64, err error) {
-	// DeleteByPrefix is not supported within a transaction context.
-	// This operation should be performed outside of a transaction.
-	return 0, 0, fmt.Errorf("DeleteByPrefix not supported within transaction context")
-}
-
-// tryFastPathCompoundQuery attempts to handle common compound query patterns
-// using structured scanning rather than regex capture arrays.
-// Returns (result, true) if handled, (nil, false) if the query should go through normal routing.
-//
-// Pattern: MATCH (a:Label), (b:Label) WITH a, b LIMIT 1 CREATE (a)-[r:Type]->(b) DELETE r
-// This is a very common pattern in benchmarks and relationship tests.
-func (e *StorageExecutor) tryFastPathCompoundQuery(ctx context.Context, cypher string) (*ExecuteResult, bool) {
-	if match, ok := matchCompoundQueryShape(cypher); ok {
-		switch match.Kind {
-		case shapeKindCompoundCreateDeleteRel:
-			e.markCompoundQueryFastPathUsed()
-			return e.executeFastPathCreateDeleteRel(
-				match.Captures.String("label1"),
-				match.Captures.String("label2"),
-				match.Captures.String("prop1"),
-				match.Captures.Any("value1"),
-				match.Captures.String("prop2"),
-				match.Captures.Any("value2"),
-				match.Captures.String("rel_type"),
-			)
-		case shapeKindCompoundPropCreateDeleteRel:
-			e.markCompoundQueryFastPathUsed()
-			return e.executeFastPathCreateDeleteRel(
-				match.Captures.String("label1"),
-				match.Captures.String("label2"),
-				match.Captures.String("prop1"),
-				match.Captures.Any("value1"),
-				match.Captures.String("prop2"),
-				match.Captures.Any("value2"),
-				match.Captures.String("rel_type"),
-			)
-		case shapeKindCompoundPropCreateDeleteReturnCountRel:
-			e.markCompoundQueryFastPathUsed()
-			return e.executeFastPathCreateDeleteRelCount(
-				match.Captures.String("label1"),
-				match.Captures.String("label2"),
-				match.Captures.String("prop1"),
-				match.Captures.Any("value1"),
-				match.Captures.String("prop2"),
-				match.Captures.Any("value2"),
-				match.Captures.String("rel_type"),
-				match.Captures.String("rel_var"),
-			)
-		}
-	}
-
-	return nil, false
-}
-
-// executeFastPathCreateDeleteRel executes the fast-path for MATCH...CREATE...DELETE patterns.
-// If prop1/prop2 are empty, uses GetFirstNodeByLabel. Otherwise uses property lookup.
-func (e *StorageExecutor) executeFastPathCreateDeleteRel(label1, label2, prop1 string, val1 any, prop2 string, val2 any, relType string) (*ExecuteResult, bool) {
-	var err error
-
-	// Get node1
-	if prop1 == "" {
-		_, err = storage.FirstNodeIDByLabel(e.storage, label1)
-	} else {
-		node1 := e.findNodeByLabelAndProperty(label1, prop1, val1)
-		if node1 == nil {
-			return nil, false
-		}
-	}
-	if err != nil {
-		return nil, false
-	}
-
-	// Get node2
-	if prop2 == "" {
-		_, err = storage.FirstNodeIDByLabel(e.storage, label2)
-	} else {
-		node2 := e.findNodeByLabelAndProperty(label2, prop2, val2)
-		if node2 == nil {
-			return nil, false
-		}
-	}
-	if err != nil {
-		return nil, false
-	}
-
-	// Optimization: This pattern creates a relationship and deletes it in the same
-	// statement without returning it. The relationship is not observable to the user,
-	// and the net graph effect is a no-op, so we skip storage writes entirely.
-	//
-	// We still validate that both endpoints exist (via the lookups above) and we
-	// still return correct query stats for Neo4j compatibility.
-
-	return &ExecuteResult{
-		Columns: []string{},
-		Rows:    [][]interface{}{},
-		Stats: &QueryStats{
-			RelationshipsCreated: 1,
-			RelationshipsDeleted: 1,
-		},
-	}, true
-}
-
-func (e *StorageExecutor) executeFastPathCreateDeleteRelCount(label1, label2, prop1 string, val1 any, prop2 string, val2 any, relType string, relVar string) (*ExecuteResult, bool) {
-	var err error
-
-	if prop1 == "" {
-		_, err = storage.FirstNodeIDByLabel(e.storage, label1)
-	} else {
-		node1 := e.findNodeByLabelAndProperty(label1, prop1, val1)
-		if node1 == nil {
-			return nil, false
-		}
-	}
-	if err != nil {
-		return nil, false
-	}
-
-	if prop2 == "" {
-		_, err = storage.FirstNodeIDByLabel(e.storage, label2)
-	} else {
-		node2 := e.findNodeByLabelAndProperty(label2, prop2, val2)
-		if node2 == nil {
-			return nil, false
-		}
-	}
-	if err != nil {
-		return nil, false
-	}
-
-	return &ExecuteResult{
-		Columns: []string{"count(" + relVar + ")"},
-		Rows:    [][]interface{}{{int64(1)}},
-		Stats: &QueryStats{
-			RelationshipsCreated: 1,
-			RelationshipsDeleted: 1,
-		},
-	}, true
-}
-
-// findNodeByLabelAndProperty finds a node by label and a single property value.
-// Uses the node lookup cache for O(1) repeated lookups.
-func (e *StorageExecutor) findNodeByLabelAndProperty(label, prop string, val any) *storage.Node {
-	e.ensureNodeLookupCache()
-
-	// Try cache first (with proper locking)
-	cacheKey := fmt.Sprintf("%s:{%s:%v}", label, prop, val)
-	cacheMu := e.nodeLookupCacheLock()
-	cacheMu.RLock()
-	if cached, ok := e.nodeLookupCache[cacheKey]; ok {
-		cacheMu.RUnlock()
-		return cached
-	}
-	cacheMu.RUnlock()
-
-	// Scan nodes with label
-	nodes, err := e.storage.GetNodesByLabel(label)
-	if err != nil {
-		return nil
-	}
-
-	// Find matching node
-	for _, node := range nodes {
-		if nodeVal, ok := node.Properties[prop]; ok {
-			if fmt.Sprintf("%v", nodeVal) == fmt.Sprintf("%v", val) {
-				// Cache for next time (with proper locking)
-				cacheMu.Lock()
-				e.nodeLookupCache[cacheKey] = node
-				cacheMu.Unlock()
-				return node
-			}
-		}
-	}
-
-	return nil
-}
-
-// isSystemCommandNoGraph returns true for statements that operate on database metadata
-// (CREATE/DROP DATABASE, SHOW DATABASES, etc.) and must not use the async engine or
-// implicit transactions. These are routed to executeWithoutTransaction directly.
-func isSystemCommandNoGraph(cypher string) bool {
-	return findMultiWordKeywordIndex(cypher, "CREATE", "COMPOSITE DATABASE") == 0 ||
-		findMultiWordKeywordIndex(cypher, "CREATE", "DATABASE") == 0 ||
-		findMultiWordKeywordIndex(cypher, "CREATE", "ALIAS") == 0 ||
-		findMultiWordKeywordIndex(cypher, "DROP", "COMPOSITE DATABASE") == 0 ||
-		findMultiWordKeywordIndex(cypher, "DROP", "DATABASE") == 0 ||
-		findMultiWordKeywordIndex(cypher, "DROP", "ALIAS") == 0 ||
-		findMultiWordKeywordIndex(cypher, "SHOW", "DATABASES") == 0 ||
-		findMultiWordKeywordIndex(cypher, "ALTER", "DATABASE") == 0
-}
-
-func isShowConstraintContractsCommand(cypher string) bool {
-	return findMultiWordKeywordIndex(cypher, "SHOW", "CONSTRAINT CONTRACTS") == 0
-}
-
-// executeWithoutTransaction executes query without transaction wrapping (original path).
-func (e *StorageExecutor) executeWithoutTransaction(ctx context.Context, cypher string, upperQuery string) (*ExecuteResult, error) {
-	// FAST PATH: Simple MATCH-return-limit reads should never be routed through
-	// heavier compound planners. Keep this check first to avoid regressions when
-	// adding complex routing rules.
-	if result, handled := e.tryFastPathSimpleMatchReturnLimit(ctx, cypher, upperQuery); handled {
-		return result, nil
-	}
-	if result, handled := e.tryFastPathAnyMatchVectorCosine(ctx, cypher, upperQuery); handled {
-		return result, nil
-	}
-
-	// FAST PATH: Check for common compound query patterns using pre-compiled regex
-	// This avoids multiple findKeywordIndex calls for frequently-used patterns
-	if result, handled := e.tryFastPathCompoundQuery(ctx, cypher); handled {
-		return result, nil
-	}
-
-	// Route to appropriate handler based on query type
-	// upperQuery is passed in to avoid redundant conversion
-
-	// Cache keyword checks to avoid repeated searches
-	startsWithMatch := strings.HasPrefix(upperQuery, "MATCH")
-	startsWithCreate := strings.HasPrefix(upperQuery, "CREATE")
-	startsWithMerge := strings.HasPrefix(upperQuery, "MERGE")
-
-	// Correlated MATCH ... CALL { ... } must route before generic MATCH...CREATE/MERGE
-	// detection, because CREATE/MERGE/SET tokens can legally appear inside the
-	// CALL subquery body and should not trigger outer compound handlers.
-	if startsWithMatch && hasSubqueryPattern(cypher, callSubqueryRe) {
-		return e.executeMatchWithCallSubquery(ctx, cypher)
-	}
-
-	// MATCH ... CALL procedure() must route before generic MATCH ... CREATE/MERGE
-	// detection because procedure names such as db.create.setNodeVectorProperty
-	// contain clause-looking tokens that are not outer query clauses.
-	if startsWithMatch {
-		callIdx := findKeywordIndex(cypher, "CALL")
-		if callIdx > 0 {
-			callPart := strings.TrimSpace(cypher[callIdx:])
-			if !isCallSubquery(callPart) {
-				prefix := cypher[:callIdx]
-				hasMutationBeforeCall := findKeywordIndexInContext(prefix, "MERGE") >= 0 ||
-					findKeywordIndexInContext(prefix, "CREATE") >= 0 ||
-					findKeywordIndexInContext(prefix, "SET") >= 0 ||
-					findKeywordIndexInContext(prefix, "DELETE") >= 0 ||
-					findKeywordIndexInContext(prefix, "DETACH DELETE") >= 0 ||
-					findKeywordIndexInContext(prefix, "REMOVE") >= 0
-				if hasMutationBeforeCall {
-					goto skipMatchCallRoute
-				}
-				if findKeywordIndex(cypher[:callIdx], "WITH") > 0 {
-					return e.executeMatchWithClause(ctx, cypher)
-				}
-				return e.executeMatchWithCallProcedure(ctx, cypher)
-			}
-		}
-	}
-skipMatchCallRoute:
-
-	// MERGE queries get special handling - they have their own ON CREATE SET / ON MATCH SET logic
-	if startsWithMerge {
-		// Complex MERGE pipelines that include OPTIONAL MATCH / WITH / WHERE should
-		// use the segment executor instead of the single-MERGE parser.
-		if findKeywordIndexInContext(cypher, "OPTIONAL MATCH") > 0 ||
-			findKeywordIndexInContext(cypher, "WITH") > 0 ||
-			findKeywordIndexInContext(cypher, "WHERE") > 0 {
-			return e.executeMultipleMerges(ctx, cypher)
-		}
-		// Check for multiple MERGEs without WITH (e.g., MERGE (a) MERGE (b) MERGE (a)-[:REL]->(b))
-		firstMergeEnd := findKeywordIndex(cypher[5:], ")")
-		if firstMergeEnd > 0 {
-			afterFirstMerge := cypher[5+firstMergeEnd+1:]
-			secondMergeIdx := findKeywordIndex(afterFirstMerge, "MERGE")
-			if secondMergeIdx >= 0 {
-				return e.executeMultipleMerges(ctx, cypher)
-			}
-		}
-		return e.executeMerge(ctx, cypher)
-	}
-
-	// Cache findKeywordIndex results for compound query detection
-	var mergeIdx, createIdx, withIdx, deleteIdx, optionalMatchIdx int = -1, -1, -1, -1, -1
-
-	if startsWithMatch {
-		// Only search for keywords if query starts with MATCH
-		mergeIdx = findKeywordIndex(cypher, "MERGE")
-		createIdx = findKeywordIndex(cypher, "CREATE")
-		optionalMatchIdx = findMultiWordKeywordIndex(cypher, "OPTIONAL", "MATCH")
-	} else if startsWithCreate {
-		// Check for multiple CREATE statements (e.g., CREATE (a) CREATE (b) CREATE (a)-[:REL]->(b))
-		firstCreateEnd := findKeywordIndex(cypher[6:], ")")
-		if firstCreateEnd > 0 {
-			afterFirstCreate := cypher[6+firstCreateEnd+1:]
-			secondCreateIdx := findKeywordIndex(afterFirstCreate, "CREATE")
-			if secondCreateIdx >= 0 {
-				return e.executeMultipleCreates(ctx, cypher)
-			}
-		}
-		// Only search for WITH/DELETE if query starts with CREATE
-		withIdx = findKeywordIndex(cypher, "WITH")
-		if withIdx > 0 {
-			deleteIdx = findKeywordIndex(cypher, "DELETE")
-		}
-	}
-
-	// Compound queries: MATCH ... MERGE ... (with variable references)
-	if startsWithMatch && mergeIdx > 0 {
-		return e.executeCompoundMatchMerge(ctx, cypher)
-	}
-
-	// Compound queries: MATCH ... CREATE ... (create relationship between matched nodes)
-	if startsWithMatch && createIdx > 0 {
-		// Try the general pipeline executor first — it handles composite
-		// queries like MATCH ... CREATE ... WITH ... UNWIND ... MATCH ...
-		// CREATE ... which executeCompoundMatchCreate miscategorises
-		// (treats everything after the first CREATE as CREATE patterns,
-		// consuming intervening WITH/UNWIND/MATCH clauses as property map
-		// garbage).
-		if result, ok, err := e.executePipeline(ctx, cypher); ok {
-			return result, err
-		}
-		return e.executeCompoundMatchCreate(ctx, cypher)
-	}
-
-	// Compound queries: CREATE ... WITH ... DELETE (create then delete in same statement)
-	if startsWithCreate && withIdx > 0 && deleteIdx > 0 {
-		return e.executeCompoundCreateWithDelete(ctx, cypher)
-	}
-
-	// UNWIND pipelines may contain trailing DELETE/SET/REMOVE keywords.
-	// Route UNWIND before generic mutation keyword checks so
-	// `UNWIND ... MATCH ... REMOVE ...` does not get misrouted to executeRemove.
-	if findKeywordIndex(cypher, "UNWIND") == 0 {
-		return e.executeUnwind(ctx, cypher)
-	}
-
-	// Cache contains checks for DELETE - use word-boundary-aware detection
-	// Note: Can't use " DELETE " because DELETE is often followed by variable name (DELETE n)
-	// findKeywordIndex handles word boundaries properly (won't match 'ToDelete' in string literals)
-	hasDelete := findKeywordIndex(cypher, "DELETE") > 0 // Must be after MATCH, not at start
-	hasDetachDelete := containsKeywordOutsideStrings(cypher, "DETACH DELETE")
-
-	// Check for compound queries - MATCH ... DELETE, MATCH ... SET, etc.
-	if hasDelete || hasDetachDelete {
-		return e.executeDelete(ctx, cypher)
-	}
-
-	// Cache SET-related checks - use string-literal-aware detection to avoid
-	// matching keywords inside user content like 'MATCH (n) SET n.x = 1'
-	// Note: findKeywordIndex already checks word boundaries, so no need for leading space
-	hasSet := containsKeywordOutsideStrings(cypher, "SET")
-	hasOnCreateSet := containsKeywordOutsideStrings(cypher, "ON CREATE SET")
-	hasOnMatchSet := containsKeywordOutsideStrings(cypher, "ON MATCH SET")
-
-	// NEO4J COMPAT: Handle CREATE ... SET pattern (e.g., CREATE (n) SET n.x = 1)
-	// Neo4j allows SET immediately after CREATE without requiring MATCH
-	if startsWithCreate && !isCreateProcedureCommand(cypher) && hasSet && !hasOnCreateSet && !hasOnMatchSet &&
-		findMultiWordKeywordIndex(cypher, "CREATE", "DECAY PROFILE") != 0 &&
-		findMultiWordKeywordIndex(cypher, "CREATE", "PROMOTION PROFILE") != 0 &&
-		findMultiWordKeywordIndex(cypher, "CREATE", "PROMOTION POLICY") != 0 {
-		return e.executeCreateSet(ctx, cypher)
-	}
-
-	// Check for ALTER DATABASE before generic SET (ALTER DATABASE SET LIMIT contains "SET")
-	if findMultiWordKeywordIndex(cypher, "ALTER", "DATABASE") == 0 {
-		return e.executeAlterDatabase(ctx, cypher)
-	}
-
-	// Only route to executeSet if it's a MATCH ... SET or standalone SET
-	if hasSet && !isCreateProcedureCommand(cypher) && !hasOnCreateSet && !hasOnMatchSet &&
-		findMultiWordKeywordIndex(cypher, "CREATE", "DECAY PROFILE") != 0 &&
-		findMultiWordKeywordIndex(cypher, "CREATE", "PROMOTION PROFILE") != 0 &&
-		findMultiWordKeywordIndex(cypher, "CREATE", "PROMOTION POLICY") != 0 &&
-		findMultiWordKeywordIndex(cypher, "ALTER", "DECAY PROFILE") != 0 &&
-		findMultiWordKeywordIndex(cypher, "ALTER", "PROMOTION PROFILE") != 0 &&
-		findMultiWordKeywordIndex(cypher, "ALTER", "PROMOTION POLICY") != 0 {
-		if startsWithMatch || findKeywordIndex(cypher, "SET") == 0 {
-			return e.executeSet(ctx, cypher)
-		}
-	}
-
-	// Handle MATCH ... REMOVE (property removal) - string-literal-aware
-	// Note: findKeywordIndex already checks word boundaries
-	if containsKeywordOutsideStrings(cypher, "REMOVE") {
-		return e.executeRemove(ctx, cypher)
-	}
-
-	// Compound queries: MATCH ... OPTIONAL MATCH ...
-	// But NOT when there's a WITH clause before OPTIONAL MATCH (that's handled by executeMatchWithOptionalMatch)
-	if startsWithMatch && optionalMatchIdx > 0 {
-		// Check if there's a WITH clause BEFORE OPTIONAL MATCH
-		// If so, route to the specialized handler that processes WITH first
-		withBeforeOptional := findKeywordIndex(cypher[:optionalMatchIdx], "WITH")
-		if withBeforeOptional > 0 {
-			// WITH comes before OPTIONAL MATCH - route to executeMatchWithOptionalMatch
-			return e.executeMatchWithOptionalMatch(ctx, cypher)
-		}
-		return e.executeCompoundMatchOptionalMatch(ctx, cypher)
-	}
-
-	switch {
-	case isCreateProcedureCommand(cypher):
-		return e.executeCreateProcedure(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "CREATE", "DECAY PROFILE") == 0,
-		findMultiWordKeywordIndex(cypher, "CREATE", "PROMOTION PROFILE") == 0,
-		findMultiWordKeywordIndex(cypher, "CREATE", "PROMOTION POLICY") == 0:
-		return e.executeKnowledgePolicyDDL(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "OPTIONAL", "MATCH") == 0:
-		// OPTIONAL MATCH must be at start (position 0) to be a standalone clause
-		// Handles flexible whitespace: "OPTIONAL MATCH", "OPTIONAL\tMATCH", "OPTIONAL\nMATCH", etc.
-		return e.executeOptionalMatch(ctx, cypher)
-	case startsWithMatch && isShortestPathQuery(cypher):
-		// Handle shortestPath() and allShortestPaths() queries.
-		// Parameter substitution must happen before parsing so the variable-
-		// binding lookup for `MATCH (start:Star {starId: $startId})` resolves
-		// to the actual property value rather than the literal `$startId`.
-		// Without this, `findNodeByPattern` matches no node, the executor
-		// falls through to AllNodes() × AllNodes(), and the BFS explodes.
-		spCypher := cypher
-		if params := getParamsFromContext(ctx); params != nil {
-			spCypher = e.substituteParams(spCypher, params)
-		}
-		query, err := e.parseShortestPathQuery(ctx, spCypher)
-		if err != nil {
-			return nil, err
-		}
-		return e.executeShortestPathQuery(ctx, query)
-	case startsWithMatch:
-		// Multi-MATCH chains have dedicated routing inside executeMatch (executeMultiMatch,
-		// executeChainedMatchWithAggregations). Keep them on that path to preserve WHERE
-		// semantics across MATCH boundaries.
-		matchCount := countKeywordOccurrences(upperQuery, "MATCH")
-		optionalMatchCount := countKeywordOccurrences(upperQuery, "OPTIONAL MATCH")
-		isMultiMatch := matchCount-optionalMatchCount > 1
-		if !isMultiMatch {
-			// Check for optimizable patterns FIRST
-			patternInfo := DetectQueryPattern(ctx, cypher)
-			if patternInfo.IsOptimizable() {
-				if result, ok := e.ExecuteOptimized(ctx, cypher, patternInfo); ok {
-					return result, nil
-				}
-				// Fall through to generic on optimization failure
-			}
-		}
-		return e.executeMatch(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "CREATE", "CONSTRAINT") == 0,
-		findMultiWordKeywordIndex(cypher, "CREATE", "RANGE INDEX") == 0,
-		findMultiWordKeywordIndex(cypher, "CREATE", "FULLTEXT INDEX") == 0,
-		findMultiWordKeywordIndex(cypher, "CREATE", "VECTOR INDEX") == 0,
-		findKeywordIndex(cypher, "CREATE INDEX") == 0:
-		// Schema commands - constraints and indexes (check more specific patterns first)
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeSchemaCommand(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "CREATE", "COMPOSITE DATABASE") == 0:
-		// System command: CREATE COMPOSITE DATABASE (check before CREATE DATABASE)
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeCreateCompositeDatabase(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "CREATE", "DATABASE") == 0:
-		// System command: CREATE DATABASE (check before generic CREATE)
-		// Must be at start (position 0) to be a standalone clause
-		// Handles flexible whitespace: "CREATE DATABASE", "CREATE\tDATABASE", "CREATE\nDATABASE", etc.
-		return e.executeCreateDatabase(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "CREATE", "ALIAS") == 0:
-		// System command: CREATE ALIAS (check before generic CREATE)
-		// Must be at start (position 0) to be a standalone clause
-		// Handles flexible whitespace: "CREATE ALIAS", "CREATE\tALIAS", "CREATE\nALIAS", etc.
-		return e.executeCreateAlias(ctx, cypher)
-	case startsWithCreate:
-		return e.executeCreate(ctx, cypher)
-	case hasDelete || hasDetachDelete:
-		// DELETE/DETACH DELETE already detected above with findKeywordIndex
-		return e.executeDelete(ctx, cypher)
-	case findKeywordIndex(cypher, "CALL") == 0:
-		// Distinguish CALL {} subquery from CALL procedure()
-		// Must be at start (position 0) to be a standalone clause
-		if isCallSubquery(cypher) {
-			return e.executeCallSubquery(ctx, cypher)
-		}
-		return e.executeCall(ctx, cypher)
-	case findKeywordIndex(cypher, "RETURN") == 0:
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeReturn(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "DROP", "COMPOSITE DATABASE") == 0:
-		// System command: DROP COMPOSITE DATABASE (check before DROP DATABASE)
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeDropCompositeDatabase(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "DROP", "DATABASE") == 0:
-		// System command: DROP DATABASE (check before generic DROP)
-		// Must be at start (position 0) to be a standalone clause
-		// Handles flexible whitespace: "DROP DATABASE", "DROP\tDATABASE", "DROP\nDATABASE", etc.
-		return e.executeDropDatabase(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "DROP", "ALIAS") == 0:
-		// System command: DROP ALIAS (check before generic DROP)
-		// Must be at start (position 0) to be a standalone clause
-		// Handles flexible whitespace: "DROP ALIAS", "DROP\tALIAS", "DROP\nALIAS", etc.
-		return e.executeDropAlias(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "DROP", "CONSTRAINT") == 0:
-		// Schema command: DROP CONSTRAINT (must not be treated as generic DROP no-op).
-		// Must be at start (position 0) to be a standalone clause.
-		return e.executeSchemaCommand(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "DROP", "DECAY PROFILE") == 0,
-		findMultiWordKeywordIndex(cypher, "DROP", "PROMOTION PROFILE") == 0,
-		findMultiWordKeywordIndex(cypher, "DROP", "PROMOTION POLICY") == 0:
-		return e.executeKnowledgePolicyDDL(ctx, cypher)
-	case isDropProcedureCommand(cypher):
-		return e.executeDropProcedure(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "DROP", "INDEX") == 0:
-		// DROP INDEX — execute real drop against the target schema.
-		return e.executeDropIndex(ctx, cypher)
-	case findKeywordIndex(cypher, "DROP") == 0:
-		// Other DROP variants (not INDEX, not CONSTRAINT, not ALIAS, not PROCEDURE)
-		// treat as no-op for forward compatibility.
-		return &ExecuteResult{Columns: []string{}, Rows: [][]interface{}{}}, nil
-	case findKeywordIndex(cypher, "WITH") == 0:
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeWith(ctx, cypher)
-	case findKeywordIndex(cypher, "UNWIND") == 0:
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeUnwind(ctx, cypher)
-	case findKeywordIndex(cypher, "UNION ALL") >= 0:
-		// UNION ALL can appear anywhere in query
-		return e.executeUnion(ctx, cypher, true)
-	case findKeywordIndex(cypher, "UNION") >= 0:
-		// UNION can appear anywhere in query
-		return e.executeUnion(ctx, cypher, false)
-	case findKeywordIndex(cypher, "FOREACH") == 0:
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeForeach(ctx, cypher)
-	case findKeywordIndex(cypher, "LOAD CSV") == 0:
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeLoadCSV(ctx, cypher)
-	// SHOW commands for Neo4j compatibility
-	case findMultiWordKeywordIndex(cypher, "SHOW", "FULLTEXT INDEXES") == 0,
-		findMultiWordKeywordIndex(cypher, "SHOW", "FULLTEXT INDEX") == 0,
-		findMultiWordKeywordIndex(cypher, "SHOW", "RANGE INDEXES") == 0,
-		findMultiWordKeywordIndex(cypher, "SHOW", "RANGE INDEX") == 0,
-		findMultiWordKeywordIndex(cypher, "SHOW", "VECTOR INDEXES") == 0,
-		findMultiWordKeywordIndex(cypher, "SHOW", "VECTOR INDEX") == 0:
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeShowIndexes(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "SHOW", "INDEXES") == 0,
-		findMultiWordKeywordIndex(cypher, "SHOW", "INDEX") == 0:
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeShowIndexes(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "SHOW", "DECAY PROFILES") == 0,
-		findMultiWordKeywordIndex(cypher, "SHOW", "PROMOTION PROFILES") == 0,
-		findMultiWordKeywordIndex(cypher, "SHOW", "PROMOTION POLICIES") == 0:
-		return e.executeKnowledgePolicyDDL(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "SHOW", "CONSTRAINTS") == 0,
-		findMultiWordKeywordIndex(cypher, "SHOW", "CONSTRAINT") == 0:
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeShowConstraints(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "SHOW", "PROCEDURES") == 0:
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeShowProcedures(ctx, cypher)
-	case findKeywordIndex(cypher, "SHOW FUNCTIONS") == 0:
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeShowFunctions(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "SHOW", "COMPOSITE DATABASES") == 0:
-		// System command: SHOW COMPOSITE DATABASES (check before SHOW DATABASES)
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeShowCompositeDatabases(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "SHOW", "CONSTITUENTS") == 0:
-		// System command: SHOW CONSTITUENTS
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeShowConstituents(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "SHOW", "DATABASES") == 0:
-		// System command: SHOW DATABASES (plural - check before singular)
-		// Must be at start (position 0) to be a standalone clause
-		// Handles flexible whitespace: "SHOW DATABASES", "SHOW\tDATABASES", "SHOW\nDATABASES", etc.
-		return e.executeShowDatabases(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "SHOW", "DATABASE") == 0:
-		// System command: SHOW DATABASE (singular)
-		// Must be at start (position 0) to be a standalone clause
-		// Handles flexible whitespace: "SHOW DATABASE", "SHOW\tDATABASE", "SHOW\nDATABASE", etc.
-		return e.executeShowDatabase(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "SHOW", "ALIASES") == 0:
-		// System command: SHOW ALIASES
-		// Must be at start (position 0) to be a standalone clause
-		// Handles flexible whitespace: "SHOW ALIASES", "SHOW\tALIASES", "SHOW\nALIASES", etc.
-		return e.executeShowAliases(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "ALTER", "COMPOSITE DATABASE") == 0:
-		// System command: ALTER COMPOSITE DATABASE (check before ALTER DATABASE)
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeAlterCompositeDatabase(ctx, cypher)
-	case findMultiWordKeywordIndex(cypher, "ALTER", "DECAY PROFILE") == 0,
-		findMultiWordKeywordIndex(cypher, "ALTER", "PROMOTION PROFILE") == 0,
-		findMultiWordKeywordIndex(cypher, "ALTER", "PROMOTION POLICY") == 0:
-		return e.executeKnowledgePolicyDDL(ctx, cypher)
-	// Note: ALTER DATABASE is handled earlier (before SET check) to avoid routing conflict
-	case findMultiWordKeywordIndex(cypher, "SHOW", "LIMITS") == 0:
-		// System command: SHOW LIMITS
-		// Must be at start (position 0) to be a standalone clause
-		return e.executeShowLimits(ctx, cypher)
-	default:
-		firstWord := strings.Split(upperQuery, " ")[0]
-		return nil, fmt.Errorf("unsupported query type: %s (supported: MATCH, CREATE, MERGE, DELETE, SET, REMOVE, RETURN, WITH, UNWIND, CALL, FOREACH, LOAD CSV, SHOW, DROP, ALTER)", firstWord)
-	}
-}
-
-// executeReturn handles simple RETURN statements (e.g., "RETURN 1").
-func (e *StorageExecutor) executeReturn(ctx context.Context, cypher string) (*ExecuteResult, error) {
-	// Substitute parameters before processing
-	if params := getParamsFromContext(ctx); params != nil {
-		cypher = e.substituteParams(cypher, params)
-	}
-
-	// Parse RETURN clause - use word boundary detection
-	returnIdx := findKeywordIndex(cypher, "RETURN")
-	if returnIdx == -1 {
-		return nil, fmt.Errorf("RETURN clause not found in query: %q", truncateQuery(cypher, 80))
-	}
-
-	returnClause := strings.TrimSpace(cypher[returnIdx+6:])
-	// Strip trailing modifiers from RETURN projection (single-row RETURN path).
-	// This prevents "ORDER BY ..." from being treated as an additional projection item.
-	if cut := firstTopLevelModifierIndex(returnClause); cut >= 0 {
-		returnClause = strings.TrimSpace(returnClause[:cut])
-	}
-
-	// Handle simple literal returns like "RETURN 1" or "RETURN true"
-	parts := splitReturnExpressions(returnClause)
-	columns := make([]string, 0, len(parts))
-	values := make([]interface{}, 0, len(parts))
-
-	for _, part := range parts {
-		part = strings.TrimSpace(part)
-
-		// Check for alias (AS)
-		alias := part
-		upperPart := strings.ToUpper(part)
-		if asIdx := strings.Index(upperPart, " AS "); asIdx != -1 {
-			alias = strings.TrimSpace(part[asIdx+4:])
-			part = strings.TrimSpace(part[:asIdx])
-		}
-
-		columns = append(columns, alias)
-
-		// Handle NULL literal explicitly first
-		if strings.EqualFold(part, "null") {
-			values = append(values, nil)
-			continue
-		}
-
-		// Fabric APPLY correlated bindings (e.g., RETURN textKey, textKey128, texts).
-		if isValidIdentifier(part) {
-			if v, ok := e.fabricRecordBindings[part]; ok {
-				values = append(values, v)
-				continue
-			}
-		}
-
-		// Try to evaluate as a function or expression first
-		result := e.evaluateExpressionWithContext(ctx, part, nil, nil)
-		if result != nil {
-			values = append(values, result)
-			continue
-		}
-
-		// Parse literal value
-		if part == "1" || strings.HasPrefix(strings.ToLower(part), "true") {
-			values = append(values, int64(1))
-		} else if part == "0" || strings.HasPrefix(strings.ToLower(part), "false") {
-			values = append(values, int64(0))
-		} else if strings.HasPrefix(part, "'") && strings.HasSuffix(part, "'") {
-			values = append(values, part[1:len(part)-1])
-		} else if strings.HasPrefix(part, "\"") && strings.HasSuffix(part, "\"") {
-			values = append(values, part[1:len(part)-1])
-		} else {
-			// Try to parse as number
-			if val, err := strconv.ParseInt(part, 10, 64); err == nil {
-				values = append(values, val)
-			} else if val, err := strconv.ParseFloat(part, 64); err == nil {
-				values = append(values, val)
-			} else {
-				// Return as string
-				values = append(values, part)
-			}
-		}
-	}
-
-	return &ExecuteResult{
-		Columns: columns,
-		Rows:    [][]interface{}{values},
-	}, nil
-}
-
-func firstTopLevelModifierIndex(clause string) int {
-	cut := -1
-	for _, kw := range []string{"ORDER BY", "SKIP", "LIMIT"} {
-		if idx := topLevelKeywordIndex(clause, kw); idx >= 0 && (cut == -1 || idx < cut) {
-			cut = idx
-		}
-	}
-	return cut
-}
-
-// splitReturnExpressions splits RETURN expressions by comma, respecting parentheses and brackets depth
-func splitReturnExpressions(clause string) []string {
-	var parts []string
-	var current strings.Builder
-	parenDepth := 0
-	bracketDepth := 0
-	inQuote := false
-	quoteChar := rune(0)
-
-	for _, ch := range clause {
-		switch {
-		case (ch == '\'' || ch == '"') && !inQuote:
-			inQuote = true
-			quoteChar = ch
-			current.WriteRune(ch)
-		case ch == quoteChar && inQuote:
-			inQuote = false
-			quoteChar = 0
-			current.WriteRune(ch)
-		case ch == '(' && !inQuote:
-			parenDepth++
-			current.WriteRune(ch)
-		case ch == ')' && !inQuote:
-			parenDepth--
-			current.WriteRune(ch)
-		case ch == '[' && !inQuote:
-			bracketDepth++
-			current.WriteRune(ch)
-		case ch == ']' && !inQuote:
-			bracketDepth--
-			current.WriteRune(ch)
-		case ch == ',' && parenDepth == 0 && bracketDepth == 0 && !inQuote:
-			parts = append(parts, current.String())
-			current.Reset()
-		default:
-			current.WriteRune(ch)
-		}
-	}
-
-	if current.Len() > 0 {
-		parts = append(parts, current.String())
-	}
-
-	return parts
-}
-
-// validateSyntax performs syntax validation.
-// When NORNICDB_PARSER=antlr, uses ANTLR for strict OpenCypher grammar validation.
-// When NORNICDB_PARSER=nornic (default), uses fast inline validation.
-func (e *StorageExecutor) validateSyntax(cypher string) error {
-	// Use ANTLR parser for validation when configured
-	if config.IsANTLRParser() {
-		return e.validateSyntaxANTLR(cypher)
-	}
-	return e.validateSyntaxNornic(cypher)
-}
-
-// validateSyntaxANTLR uses ANTLR for strict OpenCypher grammar validation.
-// Provides detailed error messages with line/column information.
-func (e *StorageExecutor) validateSyntaxANTLR(cypher string) error {
-	return antlr.Validate(cypher)
-}
-
-// validateSyntaxNornic performs fast inline syntax validation.
-func (e *StorageExecutor) validateSyntaxNornic(cypher string) error {
-	if e.hasCachedValidSyntax(cypher) {
-		return nil
-	}
-	if !hasValidStartKeyword(cypher) {
-		return fmt.Errorf("syntax error: query must start with a valid clause (MATCH, CREATE, MERGE, DELETE, CALL, SHOW, EXPLAIN, PROFILE, ALTER, USE, BEGIN, COMMIT, ROLLBACK, etc.)")
-	}
-
-	// Check balanced parentheses
-	parenCount := 0
-	bracketCount := 0
-	braceCount := 0
-	inString := false
-	stringChar := byte(0)
-
-	for i := 0; i < len(cypher); i++ {
-		c := cypher[i]
-
-		if inString {
-			if c == stringChar && (i == 0 || cypher[i-1] != '\\') {
-				inString = false
-			}
-			continue
-		}
-
-		switch c {
-		case '"', '\'':
-			inString = true
-			stringChar = c
-		case '(':
-			parenCount++
-		case ')':
-			parenCount--
-		case '[':
-			bracketCount++
-		case ']':
-			bracketCount--
-		case '{':
-			braceCount++
-		case '}':
-			braceCount--
-		}
-
-		if parenCount < 0 || bracketCount < 0 || braceCount < 0 {
-			return fmt.Errorf("syntax error: unbalanced brackets at position %d", i)
-		}
-	}
-
-	if parenCount != 0 {
-		return fmt.Errorf("syntax error: unbalanced parentheses")
-	}
-	if bracketCount != 0 {
-		return fmt.Errorf("syntax error: unbalanced square brackets")
-	}
-	if braceCount != 0 {
-		return fmt.Errorf("syntax error: unbalanced curly braces")
-	}
-	if inString {
-		return fmt.Errorf("syntax error: unclosed quote")
-	}
-
-	e.markCachedValidSyntax(cypher)
-	return nil
-}
-
-var validSyntaxStarts = [...]string{
-	"MATCH", "CREATE", "MERGE", "DELETE", "DETACH", "CALL", "RETURN", "WITH",
-	"UNWIND", "OPTIONAL", "DROP", "SHOW", "FOREACH", "LOAD", "EXPLAIN",
-	"PROFILE", "ALTER", "USE", "BEGIN", "COMMIT", "ROLLBACK",
-}
-
-func hasValidStartKeyword(cypher string) bool {
-	for _, start := range validSyntaxStarts {
-		if startsWithKeywordFold(cypher, start) {
-			return true
-		}
-	}
-	return false
-}
-
-// ensureSyntaxValidationCache lazily installs the syntax-validation cache
-// pointer using sync.Once so concurrent CALL { ... } subqueries (which fan
-// out via executeCallTailParallel) cannot race on the pointer write. The
-// underlying cache itself is already mutex-guarded; the race was on the
-// initial pointer assignment.
-func (e *StorageExecutor) ensureSyntaxValidationCache() *syntaxValidationCache {
-	e.syntaxValidationOnce.Do(func() {
-		if e.syntaxValidationCache == nil {
-			e.syntaxValidationCache = &syntaxValidationCache{
-				cache: make(map[string]struct{}, 1024),
-				max:   4096,
-			}
-		}
-	})
-	return e.syntaxValidationCache
-}
-
-func (e *StorageExecutor) hasCachedValidSyntax(cypher string) bool {
-	if cypher == "" {
-		return false
-	}
-	c := e.ensureSyntaxValidationCache()
-	c.mu.RLock()
-	_, ok := c.cache[cypher]
-	c.mu.RUnlock()
-	return ok
-}
-
-func (e *StorageExecutor) markCachedValidSyntax(cypher string) {
-	if cypher == "" {
-		return
-	}
-	c := e.ensureSyntaxValidationCache()
-	c.mu.Lock()
-	if len(c.cache) >= c.max {
-		for k := range c.cache {
-			delete(c.cache, k)
-			break
-		}
-	}
-	c.cache[cypher] = struct{}{}
-	c.mu.Unlock()
 }
